fix nasty bug that makes the sensor unreadable
Your version of the example works fine in this "hello world" format. But if you try to use this library in a real life scenario, in a program that uses other sensors, other libraries, etc,, it fails. See here:

https://forum.arduino.cc/index.php?topic=642174.0

The reason for the failure is - there  needs to be a very brief pause between colorAvailable() and readColor(). Without that pause, in more complex programs, the API call fails.

Take a look at the example that comes with the Adafruit version of the APDS9960 library. It uses the delay(), and for good reason - their code works no matter what, whereas the version you provide (without delay()) only works in this hello world example.

The change I propose implements the mandatory delay() between the two function calls. With this change, the sensor can always be read.