Add license file
This defines the license in a standardized place, which allows GitHub's automated license detection system to detect the
license type:
https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/licensing-a-repository#detecting-a-license

This is the exact license text provided by choosealicense.com
This text should not be modified in any way. Doing so may interfere with GitHub's license detection or even the licence's
function as a legal instrument.