Modernize library reference link
The previous link pointed to the bespoke library reference page, which is no longer maintained.

A new system has been implemented where the library reference content is hosted in this repository as Markdown files
under the `docs` folder, and the library reference web pages on arduino.cc are generated from that content. The new link
points to the reference that is generated by the new system.