Fix spell check false positive by ignoring word
In the latest release of the codespell tool used for automated spell checking of the files of this project, the word "totaly" was added to the codespell misspelled words dictionary as a misspelling of "totally".

This caused a false detection of the variable name `totalY` as a misspelling, resulting in a failing spell check result. Since the occurrences of this word are correct and intended in this project, the false positive is resolved by configuring codespell to ignore the problematic word.